+++ src/cli.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+import { main } from "./cli/run.js";
+
+void main().then((code) => { process.exitCode = code; });